pyproject.toml
[project]
name = "pants-basedpyright"
version = "0.3.0"
description = "Pantsbuild plugin for basedpyright type checking"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Jacob Learned"},
]
license = "MIT"
license-files = ["LICENSE"]
keywords = ["pants", "pantsbuild", "basedpyright", "monorepo", "type checking"]

[project.urls]
Homepage = "https://github.com/jacoblearned/pants-basedpyright"
Documentation = "https://github.com/jacoblearned/pants-basedpyright"
Repository = "https://github.com/jacoblearned/pants-basedpyright"
Issues = "https://github.com/jacoblearned/pants-basedpyright/issues"
Changelog = "https://github.com/jacoblearned/pants-basedpyright/blob/main/CHANGELOG.md"


[dependency-groups]
dev = [
    # We specify pantsbuild deps as dev deps so Pants doesn't try to install them
    # as transitive dependencies of this plugin
    "pantsbuild.pants>=2.27.0,<2.30.0",
    "pantsbuild.pants.testutil>=2.27.0,<2.30.0",
    "pytest>=7.0.1",
    "ruff>=0.12.11",
    "basedpyright>=1.31.4",
    "bandit>=1.8.6",
]

[build-system]
requires = ["hatchling >= 1.27"]
build-backend = "hatchling.build"

[tool.uv.sources]
"pantsbuild.pants" = {index = "pants-wheels"}
"pantsbuild.pants.testutil" = {index = "pants-wheels"}

[[tool.uv.index]]
name = "pants-wheels"
url = "https://wheels.pantsbuild.org/simple/"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "PL"]
ignore = [
    "E501",  # line too long - conflicts with fmt: https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
]

[tool.basedpyright]
pythonVersion = "3.11"
exclude = [".venv"]
reportImplicitOverride = false

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
